custom async types for Generics. Could be better, but not now.
