+++ AsyncSuffix/Analyzer/AsyncMethodNameUtil.cs
@@ -2,8 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using JetBrains.Application.Settings;
-using JetBrains.Metadata.Reader.API;
-using JetBrains.Metadata.Reader.Impl;
 using JetBrains.ReSharper.Psi;
 using JetBrains.ReSharper.Psi.CSharp.Impl;
 using JetBrains.ReSharper.Psi.CSharp.Tree;
@@ -13,28 +11,9 @@ using Sizikov.AsyncSuffix.Settings;
 
 namespace Sizikov.AsyncSuffix.Analyzer
 {
-    public sealed class AsyncMethodNameUtil
+    internal static class AsyncMethodNameUtil
     {
-        private static readonly JetHashSet<IClrTypeName> TestMethodClrAttributes = new JetHashSet<IClrTypeName>();
-
-        static AsyncMethodNameUtil()
-        {
-            TestMethodClrAttributes.Add(
-                new ClrTypeName("Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute"));
-
-            TestMethodClrAttributes.Add(
-                new ClrTypeName("Xunit.FactAttribute"));
-
-            TestMethodClrAttributes.Add(
-                new ClrTypeName("Xunit.TheoryAttribute"));
-
-            TestMethodClrAttributes.Add(
-                new ClrTypeName("NUnit.Framework.TestAttribute"));
-            TestMethodClrAttributes.Add(
-                new ClrTypeName("NUnit.Framework.TestCaseAttribute"));
-        }
-
-        public static bool IsAsyncSuffixMissing(IMethodDeclaration methodDeclaration)
+        public static bool IsAsyncSuffixMissing(this IMethodDeclaration methodDeclaration)
         {
             if (methodDeclaration.IsOverride) return false;
 
@@ -45,7 +24,7 @@ namespace Sizikov.AsyncSuffix.Analyzer
                 var excludeTestMethods = settings.GetValue(AsyncSuffixSettingsAccessor.ExcludeTestMethodsFromAnalysis);
                 if (excludeTestMethods)
                 {
-                    if (IsAnnotatedWithKnownTestAttribute(methodDeclaration))
+                    if (methodDeclaration.IsAnnotatedWithKnownTestAttribute())
                     {
                         return false;
                     }
@@ -57,40 +36,30 @@ namespace Sizikov.AsyncSuffix.Analyzer
                 }
 
                 var returnType = declaredElement.ReturnType as IDeclaredType;
-                if (returnType != null)
+                if (returnType == null) return false;
+
+                var customAsyncTypeNames = settings.EnumEntryIndices(AsyncSuffixSettingsAccessor.CustomAsyncTypes)
+                    .ToArray();
+                var customAsyncTypes = new List<IDeclaredType>();
+                customAsyncTypeNames
+                    .ForEach(type => customAsyncTypes.Add(TypeFactory.CreateTypeByCLRName(type, declaredElement.Module)));
+                var conversionRule = new CSharpTypeConversionRule(returnType.Module);
+                var isCustomAsyncType = customAsyncTypes.Any(type => returnType.IsSubtypeOf(type) || returnType.IsImplicitlyConvertibleTo(type, conversionRule));
+                if (!isCustomAsyncType)
                 {
-                    var customAsyncTypes = new List<IDeclaredType>();
-                    settings.EnumEntryIndices(AsyncSuffixSettingsAccessor.CustomAsyncTypes)
-                .ToArray()
-                .ForEach(type => customAsyncTypes.Add(TypeFactory.CreateTypeByCLRName(type, declaredElement.Module)));
-                    var conversionRule = new CSharpTypeConversionRule(returnType.Module);
-                    var isCustomAsyncType = customAsyncTypes.Any(type => returnType.IsSubtypeOf(type) || returnType.IsImplicitlyConvertibleTo(type, conversionRule));
-                   
-                    if (returnType.IsTaskType() || isCustomAsyncType)
+                    var declaredTypes = returnType.GetSuperTypes();
+                    var clrNames = declaredTypes.Select(declaredType => declaredType.GetClrName()).ToList();
+                    clrNames.Add(returnType.GetClrName());
+
+                    if (clrNames.Any(clrTypeName => customAsyncTypeNames.Contains(clrTypeName.FullName)))
                     {
-                        return true;
+                        isCustomAsyncType = true;
                     }
                 }
-            }
-            return false;
-        }
 
-        private static bool IsAnnotatedWithKnownTestAttribute(IMethodDeclaration methodDeclaration)
-        {
-            if (methodDeclaration.AttributeSectionList != null)
-            {
-                foreach (var attribute in methodDeclaration.AttributeSectionList.AttributesEnumerable)
+                if (returnType.IsTaskType() || isCustomAsyncType)
                 {
-                    var attributeClass = attribute.Name.Reference.Resolve().DeclaredElement as IClass;
-                    if (attributeClass == null)
-                    {
-                        return false;
-                    }
-                    var clrTypeName = attributeClass.GetClrName();
-                    if (TestMethodClrAttributes.Contains(clrTypeName))
-                    {
-                        return true;
-                    }
+                    return true;
                 }
             }
             return false;
